Add "mirror" GH workflows
A couple of PR workflows are "optional", but if those are enabled, those should succeed, especially for automatically merged PRs like those opened by renovate.

A PR WF that's not run is marked as "Skipped". For the "required checks" a skipped check means "failed". The [workaround described here](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks) recommends adding another WF with the same WF and step names as the "skipped one", see the description in the link for details.