New storage: Add transitional `Persist` implementation
When importing Nessie repositories, the steps to import the "incomplete"
commit objects and create the references are pretty quick, but the
mandatory import phase to create the indexes for and in each commit
object require an update operation on every imported commit. Update
operations can be quite slow on some databases.

The `Persist` implementation provided by
`:nessie-versioned-storage-transitional` delegates to two `Persist`
instances: the "intermediate" and the "eventual" one.

Initially, all operations against the transitional `Persist` are
forwarded to the "intermediate" `Persist` instance. Calling the
`transition()` method copies all data from the "intermediate" one to the
"eventual" `Persist` instance.

Ideally only `Persist` instances working on the local machine should be
used as "intermediate" `Persist` instances, for example the in-memory and
RocksDB `Persist` implementations.