CI: Replace auto-labeler with changed-files checks
The auto-labeler approach is nice, but has some disadvantages.
Each label addition triggers a new workflow run. This is usually fine
for manually added labels, but if multiple labels are added in a short
period of time, there is a non-neglectible chance that GitHub then
cancels the wrong workflow run (think: `concurrency` in `ci.yml`).

This change replaces the auto-labeler approach with an additional job in
`ci.yml` that determines the jobs to run.

For pull requests, the corresponding label is checked first, then the
list of changed files is checked against the same globs that were used
for the auto-labeling approach.