Add `renameTo` to `Conflict` response object
Clients that need to inspect conflicts returned by a merge operation need to know which keys caused a conflict, but also need to be aware of renames. This is important when clients need to retrieve the actual contents, using the right keys - the key used on the source branch (with the rename operation) and the key on the target branch (present as `Conflict.key`).

This requires the `Conflict` object to return more information, which would be a breaking change. To mitigate this, the client-spec announced by clients is bumped from `2` to `3` - the server side strips the added fields when the client-spec is `< 3`, so old clients will not receive the added fields, but new clients do.

Renames that happen(ed) on the target branch need to be inspected by the client by inspecting the diff of the merge-base and the target head.