Fix Swagger-UI in Quarkus
The change #7266 silently broke the Swagger UI in Quarkus. If Quarkus finds a `servers` section in `openapi-extras` or in an `openapi.yaml` in the classpath, it will use the `servers` from there. The `servers` added by #7266 do have an `/api/` base-path, which conflicts with the operation-paths generated by Quakus: in Nessie's `openapi.yaml` it's for example `/v2/config` whereas Quarkus generates operation paths like `/api/v2/config`. The Quarkus Swagger-UI cannot be "convinced" to use no base-path.

This change also removes the `openapi.yaml` in the "root folder" of `nessie-model.jar`, because Quarkus would use it. The "real" `openapi.yaml` is in the `nessie-model.jar` as `openapi/openapi.yaml`|`json`.